Slugignore, gitignore
Created slugignor and updated gitignore.
Folders "Fulltemplate" and "backup" won't be updated to git anymore.  These also won't be delivered to Heroku, and be removed on next commit.

They are not necessarry for the website to run and new deploys, builds will run faster by delivering smaller packages to Heroku.